+++ source/buttons/changeclubinfo.js
@@ -4,13 +4,15 @@ const { getClubDictionary, updateClub, updateList } = require('../engines/refere
 const { timeConversion } = require('../helpers.js');
 const { updateClubDetails } = require('../engines/clubEngine.js');
 const { clubEmbedBuilder } = require('../engines/messageEngine.js');
+const { SAFE_DELIMITER, SKIP_INTERACTION_HANDLING } = require('../constants.js');
 
 const mainId = "changeclubinfo";
 module.exports = new ButtonWrapper(mainId, 3000,
 	/** Set the name, description, game, image and/or color for the club with provided id */
 	(interaction, [clubId]) => {
 		const club = getClubDictionary()[clubId];
-		const modal = new ModalBuilder().setCustomId(mainId)
+		const modalCustomId = `${SKIP_INTERACTION_HANDLING}${SAFE_DELIMITER}${interaction.id}`;
+		const modal = new ModalBuilder().setCustomId(modalCustomId)
 			.setTitle("Set Club Info")
 			.addComponents(
 				new ActionRowBuilder().addComponents(
@@ -58,7 +60,7 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				)
 			);
 		interaction.showModal(modal);
-		interaction.awaitModalSubmit({ filter: interaction => interaction.customId === mainId, time: timeConversion(5, "m", "ms") }).then(async modalSubmission => {
+		interaction.awaitModalSubmit({ filter: submission => submission.customId === modalCustomId, time: timeConversion(5, "m", "ms") }).then(async modalSubmission => {
 			const { fields } = modalSubmission;
 			const errors = {};