+++ source/buttons/changeclubseats.js
@@ -4,13 +4,15 @@ const { getClubDictionary, updateClub, updateList } = require('../engines/refere
 const { clubEmbedBuilder } = require('../engines/messageEngine.js');
 const { updateClubDetails } = require('../engines/clubEngine.js');
 const { timeConversion } = require('../helpers.js');
+const { SKIP_INTERACTION_HANDLING, SAFE_DELIMITER } = require('../constants.js');
 
 const mainId = "changeclubseats";
 module.exports = new ButtonWrapper(mainId, 3000,
 	/** Set the max members and isRecruting for the club with provided id */
 	(interaction, [clubId]) => {
 		const club = getClubDictionary()[clubId];
-		const modal = new ModalBuilder().setCustomId(mainId)
+		const modalCustomId = `${SKIP_INTERACTION_HANDLING}${SAFE_DELIMITER}${interaction.id}`;
+		const modal = new ModalBuilder().setCustomId(modalCustomId)
 			.setTitle("Club Membership Settings")
 			.addComponents(
 				new ActionRowBuilder().addComponents(
@@ -23,7 +25,7 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				),
 			);
 		interaction.showModal(modal);
-		interaction.awaitModalSubmit({ filter: interaction => interaction.customId === mainId, time: timeConversion(5, "m", "ms") }).then(modalSubmission => {
+		interaction.awaitModalSubmit({ filter: submission => submission.customId === modalCustomId, time: timeConversion(5, "m", "ms") }).then(modalSubmission => {
 			const { fields } = modalSubmission;
 			const errors = {};