+++ source/util/textUtil.js
@@ -1,0 +1,17 @@
+const { commandIds } = require("../constants");
+
+/** generates a command mention, which users can click to shortcut them to using the command
+ * @param {string} fullCommand for subcommands append a whitespace and the subcommandName
+ */
+function commandMention(fullCommand) {
+	const [mainCommand] = fullCommand.split(" ");
+	if (!(mainCommand in commandIds)) {
+		return `\`/${fullCommand}\``;
+	}
+
+	return `</${fullCommand}:${commandIds[mainCommand]}>`;
+}
+
+module.exports = {
+	commandMention
+};